price type to decimal(10,2) and move cors config to spring security

+++ backend/src/main/java/com/bookstore/backend/dao/OrderDao.java
@@ -1,12 +1,14 @@
 package com.bookstore.backend.dao;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.bookstore.backend.entity.Order;
 import com.bookstore.backend.entity.OrderDetailMeta;
 
 public interface OrderDao {
-  Integer createOrder(Integer userId, String name, String phoneNumber, String address, String note, Double totalPrice);
+  Integer createOrder(Integer userId, String name, String phoneNumber, String address, String note,
+      BigDecimal totalPrice);
 
   List<Order> getOrders(Integer userId);
